Refactor: Split large discovery module source files (#1598)
Split six large files in src/discovery/ into 15 focused, single-responsibility
modules. No functional changes — all 4,372 tests pass. Original modules
re-export extracted symbols for backward compatibility.
