+++ src/discovery/DiscoveryReplayRunner.ts
@@ -1,28 +1,20 @@
-import { assertExists } from "@std/assert";
-import { getLogger } from "../utils/Logger.ts";
+/**
+ * Discovery Replay Runner Module
+ *
+ * Replays cached successful discovery changes against the current creature
+ * to find improvements without re-running the full discovery process.
+ *
+ * Helper functions extracted to:
+ * - ReplayEntryApplication.ts (applying cache entries to creatures)
+ * - ReplayHelpers.ts (evaluation, combo building, scoring utilities)
+ * - DiscoveryReplayRunnerTypes.ts (interface definitions)
+ */
+
 import type { CostName } from "../Costs.ts";
-import type { CreatureExport } from "../architecture/CreatureInterfaces.ts";
 import { CreatureUtil } from "../architecture/CreatureUtils.ts";
-import type {
-  CandidateHarmfulNeuron,
-  CandidateNeuron,
-  CandidateSquash,
-  CandidateSynapse,
-} from "../architecture/ErrorGuidedStructuralEvolution/DiscoverStructure.ts";
-import { DiscoverStructure } from "../architecture/ErrorGuidedStructuralEvolution/DiscoverStructure.ts";
-import type { RemovalCandidate } from "../architecture/ErrorGuidedStructuralEvolution/DiscoverResult.ts";
-import type {
-  CoordinatedStructuralCandidate,
-} from "../architecture/ErrorGuidedStructuralEvolution/CoordinatedStructuralCandidate.ts";
-import {
-  applyCoordinatedStructuralCandidate,
-} from "../architecture/ErrorGuidedStructuralEvolution/ApplyCoordinatedStructuralCandidate.ts";
-import { calculate as calculateScore } from "../architecture/Score.ts";
 import { createNeatConfig } from "../config/NeatConfig.ts";
-import type { NeatOptions } from "../config/NeatOptions.ts";
 import type { Creature } from "../Creature.ts";
-import { WorkerHandler } from "../multithreading/workers/WorkerHandler.ts";
-import { formatWeight } from "./FailureCache.ts";
+import type { WorkerHandler } from "../multithreading/workers/WorkerHandler.ts";
 import {
   calculatePriority,
   dequeueAll,
@@ -36,587 +28,37 @@ import {
   listSuccessEntriesSync,
   type SuccessCacheEntry,
 } from "./SuccessCache.ts";
+import { applyEntryUsingRustRequest } from "./ReplayEntryApplication.ts";
+import { isAlreadyApplied } from "./ReplayEntryApplication.ts";
+import {
+  buildComboIndices,
+  describeCombo,
+  evaluateAll,
+  type EvaluationTask,
+  mapConcurrent,
+  parseClaimedTagNumber,
+  scoreMeaningfullyDifferent,
+  setupWorkerPool,
+} from "./ReplayHelpers.ts";
+import type {
+  DiscoveryReplayDiagnostics,
+  DiscoveryReplayDirInput,
+  DiscoveryReplayDirResult,
+  DiscoveryReplayEvaluationSummary,
+  DiscoveryReplayRunnerDeps,
+  DiscoveryReplayRunnerLike,
+} from "./DiscoveryReplayRunnerTypes.ts";
 
-export interface DiscoveryReplayDirInput {
-  creature: Creature;
-  dataDir: string;
-  options: NeatOptions;
-  /**
-   * Optional timeout in minutes for the replay operation.
-   * If provided and > 0, replay will abort gracefully when the timeout is reached.
-   * This prevents replay from hanging the evolution process when time runs out.
-   */
-  timeoutMinutes?: number;
-}
-
-export interface DiscoveryReplayEvaluationSummary {
-  kind: "original" | "single" | "combo";
-  key?: string;
-  changeType?: string;
-  description?: string;
-  score: number;
-  error: number;
-  scoreDelta?: number;
-  improved: boolean;
-}
-
-export interface DiscoveryReplayDiagnostics {
-  /**
-   * Timing breakdown in milliseconds. Values are best-effort and intended for
-   * visibility (where time is being spent), not micro-benchmarking.
-   */
-  timingsMS: {
-    total?: number;
-    setupWorkers?: number;
-    listEntries?: number;
-    sortEntries?: number;
-    applySingles?: number;
-    evaluateBaselineAndSingles?: number;
-    buildCombos?: number;
-    evaluateCombos?: number;
-    selectBest?: number;
-    teardownWorkers?: number;
-  };
-  counts: {
-    workerCount: number;
-    entriesLoaded: number;
-    singlesApplied: number;
-    singlesEvaluated: number;
-    combosBuilt: number;
-    combosEvaluated: number;
-    pruned: number;
-    skippedAlreadyApplied: number;
-    skippedNotApplicable: number;
-  };
-}
-
-export interface DiscoveryReplayDirResult {
-  original: {
-    error: number;
-    score: number;
-  };
-  diagnostics?: DiscoveryReplayDiagnostics;
-  baselineRescore?: {
-    claimedScore?: number;
-    actualScore: number;
-    actualError: number;
-    changed: boolean;
-    reason: string;
-  };
-  improvement?: {
-    key?: string;
-    changeType: string;
-    error: number;
-    score: number;
-    scoreDelta: number;
-    message: string;
-    creature: unknown;
-  };
-  verifiedImprovement?: {
-    score: number;
-    error: number;
-    scoreDelta: number;
-    improved: boolean;
-    message: string;
-    creature: CreatureExport;
-  };
-  evaluatedSingles: number;
-  evaluatedCombos: number;
-  pruned: number;
-  skippedAlreadyApplied: number;
-  skippedNotApplicable: number;
-  evaluations?: DiscoveryReplayEvaluationSummary[];
-  /**
-   * True if the replay operation timed out before completing all evaluations.
-   * When timed out, the result contains partial data from evaluations completed
-   * before the timeout.
-   */
-  timedOut?: boolean;
-}
-
-export interface DiscoveryReplayRunnerLike {
-  replayDir(input: DiscoveryReplayDirInput): Promise<DiscoveryReplayDirResult>;
-}
-
-export interface DiscoveryReplayRunnerDeps {
-  listEntries?: (dir: string) => SuccessCacheEntry[];
-  /**
-   * Archive an obsolete success cache entry instead of deleting it.
-   * This preserves history of candidates that once improved but no longer do.
-   */
-  archiveEntry?: (dir: string, entry: SuccessCacheEntry) => void;
-  applyEntry?: (
-    baseCreature: Creature,
-    entry: SuccessCacheEntry,
-  ) => Creature | undefined;
-  evaluateError?: (
-    creature: Creature,
-    feedbackLoop: boolean,
-    costOfGrowth: number,
-  ) => Promise<{ error: number; score: number }>;
-}
-
-type EvaluationTask = {
-  kind: "original" | "single" | "combo";
-  creature: Creature;
-  entry?: SuccessCacheEntry;
-  description?: string;
-};
-
-/**
- * Evaluate a batch of replay tasks across a set of workers.
- *
- * @internal
- */
-export async function evaluateAll(
-  workers: WorkerHandler[],
-  tasks: EvaluationTask[],
-  feedbackLoop: boolean,
-  costOfGrowth: number,
-): Promise<Array<EvaluationTask & { error: number; score: number }>> {
-  const queue = tasks.map((task, index) => ({ task, index }));
-  const results: Array<
-    (EvaluationTask & { error: number; score: number }) | undefined
-  > = new Array(tasks.length);
-
-  const processNext = async (worker: WorkerHandler): Promise<void> => {
-    const next = queue.shift();
-    if (!next) return;
-    const { task, index } = next;
-
-    const response = await worker.evaluate(task.creature, feedbackLoop);
-    assertExists(
-      response.evaluate,
-      "Worker did not return evaluation data.",
-    );
-    const error = response.evaluate.error;
-    const score = calculateScore(task.creature, error, costOfGrowth);
-    results[index] = { ...task, error, score };
-
-    await processNext(worker);
-  };
-
-  await Promise.all(workers.map((worker) => processNext(worker)));
-  return results.filter((r) => r !== undefined);
-}
-
-function isRemovalType(changeType: string): boolean {
-  return changeType === "remove-low-impact" || changeType === "remove-neuron" ||
-    changeType === "remove-synapse";
-}
-
-function safeNumber(value: unknown): number | undefined {
-  return typeof value === "number" && Number.isFinite(value)
-    ? value
-    : undefined;
-}
-
-function parseClaimedTagNumber(
-  tags: Array<{ name: string; value?: string }> | undefined,
-  key: "score" | "error",
-): number | undefined {
-  if (!tags || tags.length === 0) return undefined;
-  for (const tag of tags) {
-    if (!tag?.name) continue;
-    if (tag.name.toLowerCase() !== key) continue;
-    const raw = tag.value ?? "";
-    const parsed = Number.parseFloat(raw);
-    if (Number.isFinite(parsed)) return parsed;
-  }
-  return undefined;
-}
-
-/**
- * Whether a claimed score (e.g. from tags) differs meaningfully from a re-computed score.
- * Uses a relaxed relative threshold so we do not report drift from normal floating-point
- * and string round-trip noise.
- *
- * Post-WASM migration, small differences are expected because: (1) WASM activation uses
- * f32 in Rust, so results can differ at the last few bits from pre-migration f64 JS or
- * between main-thread and worker WASM instances; (2) worker evaluation rebuilds the
- * creature from JSON, so weight/bias round-trip can introduce sub-ulp differences;
- * (3) the claimed score is stored via number.toString(), which can round when parsed back.
- */
-function scoreMeaningfullyDifferent(
-  claimed: number | undefined,
-  actual: number,
-): boolean {
-  if (claimed === undefined) return false;
-  if (!Number.isFinite(actual)) return true;
-  const absDiff = Math.abs(claimed - actual);
-  if (absDiff <= 1e-12) return false;
-  const denom = Math.max(1, Math.abs(claimed), Math.abs(actual));
-  return absDiff / denom > 1e-6;
-}
-
-async function mapConcurrent<TIn, TOut>(
-  inputs: readonly TIn[],
-  concurrency: number,
-  fn: (value: TIn, index: number) => Promise<TOut>,
-): Promise<TOut[]> {
-  const limit = Math.max(1, Math.floor(concurrency));
-  const results: TOut[] = new Array(inputs.length);
-  let nextIndex = 0;
-
-  const runOne = async (): Promise<void> => {
-    while (true) {
-      const i = nextIndex++;
-      if (i >= inputs.length) return;
-      // deno-lint-ignore no-await-in-loop -- This loop is the concurrency limiter; awaiting here is intentional.
-      results[i] = await fn(inputs[i], i);
-    }
-  };
-
-  const runners = Array.from(
-    { length: Math.min(limit, inputs.length) },
-    () => runOne(),
-  );
-  await Promise.all(runners);
-  return results;
-}
-
-function getRustRequest(entry: SuccessCacheEntry): Record<string, unknown> {
-  return (entry.rustRequest as Record<string, unknown>) ?? {};
-}
-
-function isNeuronPresent(creature: Creature, uuid: string): boolean {
-  return creature.neurons.some((n) => n.uuid === uuid);
-}
-
-function isSynapsePresent(
-  creature: Creature,
-  fromUUID: string,
-  toUUID: string,
-): boolean {
-  const exported = creature.exportJSON();
-  return exported.synapses.some((s) =>
-    s.fromUUID === fromUUID && s.toUUID === toUUID
-  );
-}
-
-function nearlyEqual(a: number, b: number): boolean {
-  if (a === b) return true;
-  if (!Number.isFinite(a) || !Number.isFinite(b)) return false;
-  const diff = Math.abs(a - b);
-  if (diff <= 1e-12) return true;
-  const scale = Math.max(1, Math.abs(a), Math.abs(b));
-  return diff <= 1e-7 * scale;
-}
-
-function isAlreadyApplied(
-  creature: Creature,
-  entry: SuccessCacheEntry,
-): boolean {
-  const type = entry.changeType;
-  const req = getRustRequest(entry);
-
-  // Fast-path: structural removals.
-  if (type === "remove-low-impact") {
-    const c = req.removalCandidate as RemovalCandidate | undefined;
-    return c?.neuronUUID ? !isNeuronPresent(creature, c.neuronUUID) : false;
-  }
-  if (type === "remove-neuron") {
-    const c = req.harmfulNeuronCandidate as CandidateHarmfulNeuron | undefined;
-    return c?.neuronUUID ? !isNeuronPresent(creature, c.neuronUUID) : false;
-  }
-  if (type === "remove-synapse") {
-    const c = req.harmfulSynapseCandidate as CandidateSynapse | undefined;
-    if (c?.fromNeuronUUID && c?.toNeuronUUID) {
-      return !isSynapsePresent(creature, c.fromNeuronUUID, c.toNeuronUUID);
-    }
-    const details = req.synapseDetails as {
-      fromNeuronUUID?: string;
-      toNeuronUUID?: string;
-    } | undefined;
-    if (details?.fromNeuronUUID && details?.toNeuronUUID) {
-      return !isSynapsePresent(
-        creature,
-        details.fromNeuronUUID,
-        details.toNeuronUUID,
-      );
-    }
-    return false;
-  }
-
-  if (type === "add-synapses") {
-    const c = req.synapseCandidate as CandidateSynapse | undefined;
-    if (!c?.fromNeuronUUID || !c?.toNeuronUUID) return false;
-    return isSynapsePresent(creature, c.fromNeuronUUID, c.toNeuronUUID);
-  }
-
-  if (type === "change-squash") {
-    const c = req.squashCandidate as CandidateSquash | undefined;
-    if (!c?.neuronUUID || !c?.squash) return false;
-    const neuron = creature.neurons.find((n) => n.uuid === c.neuronUUID);
-    return neuron?.squash === c.squash;
-  }
-
-  // Coordinated structural candidates are multi-op groups and may not be
-  // trivially “already applied” (eg remove/remove/add sequences).
-  // We treat them as already applied only when we can verify every operation.
-  if (type === "coordinated-structural") {
-    const spec = req.coordinatedStructuralCandidate as
-      | CoordinatedStructuralCandidate
-      | undefined;
-    const ops = spec?.operations;
-    if (!Array.isArray(ops) || ops.length === 0) return false;
-
-    // Reduce the ordered operation list into a final expected state per edge.
-    // This handles weight adjustments that Rust expresses as remove+add on the
-    // same synapse: the last op wins.
-    const expectedByEdge = new Map<
-      string,
-      { present: boolean; weight?: number }
-    >();
-    const edgeKey = (fromUUID: string, toUUID: string): string =>
-      `${fromUUID}\0${toUUID}`;
-
-    for (const op of ops) {
-      if (!op || typeof op.type !== "string") return false;
-      if (op.type === "removeSynapse") {
-        expectedByEdge.set(edgeKey(op.fromNeuronUuid, op.toNeuronUuid), {
-          present: false,
-        });
-        continue;
-      }
-      if (op.type === "addSynapse") {
-        expectedByEdge.set(edgeKey(op.fromNeuronUuid, op.toNeuronUuid), {
-          present: true,
-          weight: op.weight,
-        });
-        continue;
-      }
-      if (op.type === "setWeight") {
-        const key = edgeKey(op.fromNeuronUuid, op.toNeuronUuid);
-        const existing = expectedByEdge.get(key);
-        // setWeight is a no-op if the synapse doesn't exist. If the edge is
-        // already marked as absent (from a prior removeSynapse), leave it absent.
-        if (existing?.present === false) {
-          // No-op: leave the edge as absent.
-          continue;
-        }
-        // Otherwise, update the weight (synapse must be present).
-        expectedByEdge.set(key, {
-          present: true,
-          weight: op.weight,
-        });
-        continue;
-      }
-      // Unknown op: do not assume applied.
-      return false;
-    }
-
-    const exported = creature.exportJSON();
-    for (const [key, expected] of expectedByEdge.entries()) {
-      const [fromUUID, toUUID] = key.split("\0");
-      const synapse = exported.synapses.find((s) =>
-        s.fromUUID === fromUUID && s.toUUID === toUUID
-      );
-
-      if (!expected.present) {
-        if (synapse) return false;
-        continue;
-      }
-
-      if (!synapse) return false;
-      if (expected.weight === undefined) return false;
-      if (!nearlyEqual(synapse.weight, expected.weight)) return false;
-    }
-
-    return true;
-  }
-
-  if (type === "add-neurons") {
-    // Approximate match using the same exponent-bucketing approach as the cache key.
-    const details = req.neuronDetails as {
-      fromNeuronUUID?: string;
-      toNeuronUUID?: string;
-      incomingWeight?: number;
-      outgoingWeight?: number;
-      bias?: number;
-      squash?: string;
-    } | undefined;
-    const fromUUID = details?.fromNeuronUUID;
-    const toUUID = details?.toNeuronUUID;
-    if (!fromUUID || !toUUID) return false;
-
-    const inW = safeNumber(details?.incomingWeight);
-    const outW = safeNumber(details?.outgoingWeight);
-    const bias = safeNumber(details?.bias);
-    const squash = typeof details?.squash === "string"
-      ? details.squash
-      : undefined;
-    if (
-      inW === undefined || outW === undefined || bias === undefined || !squash
-    ) {
-      return false;
-    }
-
-    const inExp = formatWeight(inW);
-    const outExp = formatWeight(outW);
-    const biasExp = formatWeight(bias);
-
-    // Find a hidden neuron with matching squash + bias magnitude that links from->hidden->to
-    const exported = creature.exportJSON();
-    for (const neuron of creature.neurons) {
-      if (neuron.type !== "hidden") continue;
-      if (neuron.squash !== squash) continue;
-      if (formatWeight(neuron.bias) !== biasExp) continue;
-
-      const inSyn = exported.synapses.find((s) =>
-        s.fromUUID === fromUUID && s.toUUID === neuron.uuid
-      );
-      const outSyn = exported.synapses.find((s) =>
-        s.fromUUID === neuron.uuid && s.toUUID === toUUID
-      );
-      if (!inSyn || !outSyn) continue;
-      if (formatWeight(inSyn.weight) !== inExp) continue;
-      if (formatWeight(outSyn.weight) !== outExp) continue;
-      return true;
-    }
-    return false;
-  }
-
-  // Unknown type: do not assume applied.
-  return false;
-}
-
-function applyEntryUsingRustRequest(
-  baseCreature: Creature,
-  entry: SuccessCacheEntry,
-): Creature | undefined {
-  const req = getRustRequest(entry);
-  const discoveryID = entry.key || "replay";
-
-  switch (entry.changeType) {
-    case "coordinated-structural": {
-      const coordinated = req.coordinatedStructuralCandidate as
-        | CoordinatedStructuralCandidate
-        | undefined;
-      if (!coordinated) return undefined;
-      return applyCoordinatedStructuralCandidate(baseCreature, coordinated);
-    }
-    case "add-synapses": {
-      const synapse = req.synapseCandidate as CandidateSynapse | undefined;
-      if (!synapse) return undefined;
-      return DiscoverStructure.addHelpfulSynapses(
-        discoveryID,
-        baseCreature,
-        [synapse],
-      );
-    }
-    case "add-neurons": {
-      const neuron = req.neuronCandidate as CandidateNeuron | undefined;
-      if (!neuron) return undefined;
-      return DiscoverStructure.addHelpfulNeurons(
-        discoveryID,
-        baseCreature,
-        [neuron],
-      );
-    }
-    case "change-squash": {
-      const squash = req.squashCandidate as CandidateSquash | undefined;
-      if (!squash) return undefined;
-      return DiscoverStructure.changeSquash(
-        discoveryID,
-        baseCreature,
-        [squash],
-      );
-    }
-    case "remove-synapse": {
-      const synapse =
-        (req.harmfulSynapseCandidate as CandidateSynapse | undefined) ??
-          (req.synapseCandidate as CandidateSynapse | undefined);
-      const details = req.synapseDetails as
-        | { fromNeuronUUID?: string; toNeuronUUID?: string }
-        | undefined;
-      const resolved = synapse ??
-        (details?.fromNeuronUUID && details?.toNeuronUUID
-          ? {
-            fromNeuronUUID: details.fromNeuronUUID,
-            toNeuronUUID: details.toNeuronUUID,
-            weight: 0,
-            targetNeuronImpact: 0,
-            expectedCreatureErrorReduction: 0,
-            expectedCreatureScoreGain: 0,
-            improvedCount: 0,
-            totalCount: 0,
-          } satisfies CandidateSynapse
-          : undefined);
-      if (!resolved) return undefined;
-      const removed = DiscoverStructure.removeSynapse(
-        discoveryID,
-        baseCreature,
-        resolved,
-      );
-      return removed ?? undefined;
-    }
-    case "remove-neuron": {
-      const neuron = req.harmfulNeuronCandidate as
-        | CandidateHarmfulNeuron
-        | undefined;
-      if (!neuron) return undefined;
-      const removed = DiscoverStructure.removeHarmfulNeuron(
-        discoveryID,
-        baseCreature,
-        neuron,
-      );
-      return removed ?? undefined;
-    }
-    case "remove-low-impact": {
-      const c = req.removalCandidate as RemovalCandidate | undefined;
-      if (!c) return undefined;
-      const removed = DiscoverStructure.removeLowImpactNeuron(
-        discoveryID,
-        baseCreature,
-        c,
-      );
-      return removed ?? undefined;
-    }
-    default:
-      return undefined;
-  }
-}
-
-function buildComboIndices(
-  successfulSingles: Array<{ entry: SuccessCacheEntry }>,
-  maxPairwise: number,
-  maxTriples: number,
-): number[][] {
-  const combos: number[][] = [];
-  const count = successfulSingles.length;
-  if (count < 2) return combos;
-
-  // All.
-  combos.push(Array.from({ length: count }, (_, i) => i));
-
-  // All removals as a separate candidate outcome (when there are 2+ removals).
-  const removalOnly = successfulSingles
-    .map((s, i) => (isRemovalType(s.entry.changeType) ? i : -1))
-    .filter((i) => i >= 0);
-  if (removalOnly.length >= 2 && removalOnly.length < count) {
-    combos.push(removalOnly);
-  }
-
-  // Replay intentionally only evaluates the all-successful combination.
-  // Pairwise/triple exploration can be reintroduced later if needed.
-  void maxPairwise;
-  void maxTriples;
-  return combos;
-}
-
-function describeCombo(entries: SuccessCacheEntry[]): string {
-  const types = Array.from(new Set(entries.map((e) => e.changeType)));
-  const removalOnly = types.every(isRemovalType);
-  if (removalOnly) {
-    return `✂️ Replayed ${entries.length} cached pruning change${
-      entries.length === 1 ? "" : "s"
-    }`;
-  }
-  return `🏆 Replayed ${entries.length} cached change${
-    entries.length === 1 ? "" : "s"
-  } (${types.join(", ")})`;
-}
+// Re-export for backward compatibility.
+export type {
+  DiscoveryReplayDiagnostics,
+  DiscoveryReplayDirInput,
+  DiscoveryReplayDirResult,
+  DiscoveryReplayEvaluationSummary,
+  DiscoveryReplayRunnerDeps,
+  DiscoveryReplayRunnerLike,
+} from "./DiscoveryReplayRunnerTypes.ts";
+export { evaluateAll } from "./ReplayHelpers.ts";
 
 export class DiscoveryReplayRunner implements DiscoveryReplayRunnerLike {
   #deps: {
@@ -686,49 +128,15 @@ export class DiscoveryReplayRunner implements DiscoveryReplayRunnerLike {
 
     try {
       if (!this.#deps.evaluateError) {
-        const workerCount = replayConcurrency;
         const setupStart = diagnosticsEnabled ? performance.now() : 0;
-        for (let i = 0; i < workerCount; i++) {
-          const preferDirect = workerCount === 1;
-          // Issue #1567: Propagate WASM cache limits to worker threads.
-          let w = new WorkerHandler(
-            dataDir,
-            config.costName as CostName,
-            preferDirect,
-            config.customCost,
-            config.wasmCache,
-          );
-          try {
-            // Warm worker sequentially to avoid cold-cache download storms and
-            // to ensure only usable workers are added to the pool.
-            // deno-lint-ignore no-await-in-loop
-            await w.waitUntilReady();
-          } catch (err) {
-            try {
-              w.terminate();
-            } catch {
-              // Ignore termination errors.
-            }
-            if (!preferDirect) {
-              getLogger().warn(
-                "[DiscoveryReplayRunner] Worker init failed; falling back to direct execution for this worker slot.",
-                err,
-              );
-              w = new WorkerHandler(
-                dataDir,
-                config.costName as CostName,
-                true,
-                config.customCost,
-                config.wasmCache,
-              );
-              // deno-lint-ignore no-await-in-loop
-              await w.waitUntilReady();
-            } else {
-              throw err;
-            }
-          }
-          workers.push(w);
-        }
+        const pool = await setupWorkerPool({
+          workerCount: replayConcurrency,
+          dataDir,
+          costName: config.costName as CostName,
+          customCost: config.customCost,
+          wasmCache: config.wasmCache,
+        });
+        workers.push(...pool);
         if (timingsMS) timingsMS.setupWorkers = performance.now() - setupStart;
       }
